example_portfolio/requirements.txt
cmsplugin-filer==1.1.3
dj-database-url==0.4.2
Django==1.8.18
django-appconf==1.0.2
django-classy-tags==0.8.0
django-cms==3.4.4
django-filer==1.2.8
django-formtools==2.0
django-mptt==0.8.7
django-polymorphic==1.0.2
django-sekizai==0.10.0
Django-Select2==5.11.1
django-taggit==0.22.1
django-treebeard==4.1.2
djangocms-admin-style==1.2.7
djangocms-attributes-field==0.3.0
djangocms-column==1.7.0
djangocms-googlemap==1.1.1
djangocms-installer==0.9.7
djangocms-link==2.1.2
djangocms-snippet==1.9.2
djangocms-style==2.0.2
djangocms-text-ckeditor==3.5.0
djangocms-video==2.0.3
easy-thumbnails==2.4.1
html5lib==0.9999999
olefile==0.44
Pillow==4.2.1
pytz==2017.2
six==1.10.0
tzlocal==1.4
Unidecode==0.4.21
